% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_summary.R
\docType{data}
\name{buffer_summary}
\alias{buffer_summary}
\title{Monthly summaries of NDVI for TF 1km m buffer areas from 2019 - 2023}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6322 rows and 5 columns.
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/sentinel-2}
}
\usage{
buffer_summary
}
\description{
Mean, median and sd of Normalised Vegetation Index (NDVI) of the 1k buffer around each Tiny Forest location
Calculated from the NIR and Red bands of Sentinel 2 multispectral satellite images
extracted via Google Earth Engine
}
\details{
\ref{Gorelick, N. et al. (2017) ‘Google Earth Engine: Planetary-scale geospatial analysis for everyone’, Remote sensing of environment, 202, pp. 18–27. doi:10.1016/j.rse.2017.06.031.}
}
\keyword{datasets}
