% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_tf.R
\docType{data}
\name{geo_tf}
\alias{geo_tf}
\title{OS landcover for tiny forests n = 178}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1498129 rows and 11 columns.
}
\usage{
geo_tf
}
\description{
OS landcover for tiny forests n = 178
}
\keyword{datasets}
