% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_mean_ndvi_tf.R,
%   R/seasonal_rainfall_tf.R
\docType{data}
\name{seasonal_rainfall_tf}
\alias{seasonal_rainfall_tf}
\title{Quarterly rainfall for  each Tiny Forest from 2010}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9408 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9408 rows and 3 columns.
}
\source{
\url{https://catalogue.ceda.ac.uk/uuid/15deeb29cdcd4524b07560e5aad45ded}

\url{https://catalogue.ceda.ac.uk/uuid/15deeb29cdcd4524b07560e5aad45ded}
}
\usage{
seasonal_rainfall_tf

seasonal_rainfall_tf
}
\description{
Derived from CEDA 1km gridded rainfall data assigned to and averaged over each Tiny Forest buffer

Derived from CEDA 1km gridded rainfall data assigned to and averaged over each Tiny Forest buffer
}
\details{
\ref{Met Office (2022): MIDAS Open: UK daily rainfall data, v202207. NERC EDS Centre for Environmental Data Analysis, 22 September 2022. doi:10.5285/15deeb29cdcd4524b07560e5aad45ded. https://dx.doi.org/10.5285/15deeb29cdcd4524b07560e5aad45ded}

\ref{Met Office (2022): MIDAS Open: UK daily rainfall data, v202207. NERC EDS Centre for Environmental Data Analysis, 22 September 2022. doi:10.5285/15deeb29cdcd4524b07560e5aad45ded. https://dx.doi.org/10.5285/15deeb29cdcd4524b07560e5aad45ded}
}
\keyword{datasets}
