% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_dw_counts.R
\docType{data}
\name{tf_dw_counts}
\alias{tf_dw_counts}
\title{Pixel land cover counts per Tiny Forest}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 950 rows and 4 columns.
}
\usage{
tf_dw_counts
}
\description{
Pixel land cover counts per Tiny Forest
}
\keyword{datasets}
