% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_ebird_counts.R
\docType{data}
\name{tf_ebird_counts}
\alias{tf_ebird_counts}
\title{Monthly species counts of bird observations submitted to ebird by year and Tiny Forest}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 27583 rows and 4 columns.
}
\source{
\url{https://ebird.org/data/download}
}
\usage{
tf_ebird_counts
}
\description{
Counts by year and month of ebird observations in the 1km buffer around each Tiny Forest
}
\details{
\item{‘eBird’ (2020). doi:10.2173/ebird.}
}
\keyword{datasets}
