% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_latest_sf_uk.R
\docType{data}
\name{tf_latest_sf_uk}
\alias{tf_latest_sf_uk}
\title{Core data for UK Tiny Forests in a simple features format (sf)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 196 rows and 8 columns.
}
\source{
\url{https://tinyforest.earthwatch.org.uk/tiny-forest-sites}
}
\usage{
tf_latest_sf_uk
}
\description{
Tiny Forests are small scale densely planted wooded areas stocked with
native trees and understory plants. They rarely exceed 200m2 and are planted in urban/ semi-urban areas
}
\keyword{datasets}
