% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_lc.R
\docType{data}
\name{tf_lc}
\alias{tf_lc}
\title{Landvoer and landuse for TFs}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2740 rows and 4 columns.
}
\usage{
tf_lc
}
\description{
Landvoer and landuse for TFs
}
\keyword{datasets}
