% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_nbn_df_2010.R
\docType{data}
\name{tf_nbn_df_2010}
\alias{tf_nbn_df_2010}
\title{NBN Atlas data for each Tiny Forest within a buffer of 1km from 2010}
\format{
An object of class \code{data.frame} with 522767 rows and 15 columns.
}
\source{
\url{https://nbnatlas.org/}
}
\usage{
tf_nbn_df_2010
}
\description{
National Biodiversity Network (NBN) data is a mix of citizen science and formal data collated from a wide range of sources.
The data set is assembled from NBN within a 1km buffer around each TF point location
}
\keyword{datasets}
