% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_ndvi_maps.R
\docType{data}
\name{tf_ndvi_maps}
\alias{tf_ndvi_maps}
\title{Interactive maps show location of TF and variation in NDVI in 1km TF buffer}
\format{
An object of class \code{list} of length 196.
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/sentinel-2}
}
\usage{
tf_ndvi_maps
}
\description{
Using Sentinal-2 data drawn from Google Earth Engine
}
\details{
\ref{Gorelick, N. et al. (2017) ‘Google Earth Engine: Planetary-scale geospatial analysis for everyone’, Remote sensing of environment, 202, pp. 18–27. doi:10.1016/j.rse.2017.06.031.}
}
\keyword{datasets}
