% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_urban_classification.R
\docType{data}
\name{tf_urban_classification}
\alias{tf_urban_classification}
\title{Urban-rural classifiction for each Tiny Forest location}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 188 rows and 8 columns.
}
\source{
\url{https://maps.gov.scot/ATOM/shapefiles/SG_UrbanRural_2020.zip}

\url{https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/1011667/Rural_Urban_Classification_2011_lookup_tables_for_small_area_geographies.ods}
}
\usage{
tf_urban_classification
}
\description{
Each Tiny Forest was assigned to an ONS urban-rural classification by mapping the location to
a lower super output area (LSOA) and matching to the LSOA urban-rural classification database provided by ONS.
For English TFs the 10 fold classification system was used. Scotland has a slightly different system with 8 classes
}
\keyword{datasets}
