% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_water.R
\docType{data}
\name{tf_water}
\alias{tf_water}
\title{Water features for each TF buffer (only includes those with any water)}
\format{
An object of class \code{list} of length 169.
}
\usage{
tf_water
}
\description{
Water features for each TF buffer (only includes those with any water)
}
\keyword{datasets}
