% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyForestData-package.R
\docType{package}
\name{tinyForestData-package}
\alias{tinyForestData}
\alias{tinyForestData-package}
\title{tinyForestData: Data Package With Tiny Forest and Biodiversity Data}
\description{
Data relating to Tiny Forest locations, structure, biodiversity, habitat and land cover.
}
\author{
\strong{Maintainer}: Julian Flowers \email{julian.flowers13@gmail.com} (\href{https://ordic.org/0000-0001-5262-4526}{ORCID})

}
\keyword{internal}
