% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_tf.R, R/tf_landuse.R
\docType{data}
\name{geo_tf}
\alias{geo_tf}
\title{OS landcover for tiny forests n = 178}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1498129 rows and 11 columns.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1498129 rows and 11 columns.
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/sentinel-2}
}
\usage{
geo_tf

geo_tf
}
\description{
Mean, median and sd of Normalised Vegetation Index (NDVI) of the 50m buffer around each Tiny Forest location
Calculated from the NIR and Red bands of Sentinel 2 multispectral satellite images
extracted via Google Earth Engine
}
\details{
\ref{Gorelick, N. et al. (2017) ‘Google Earth Engine: Planetary-scale geospatial analysis for everyone’, Remote sensing of environment, 202, pp. 18–27. doi:10.1016/j.rse.2017.06.031.}
}
\keyword{datasets}
